requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
openpyxl>=3.1.0
XlsxWriter>=3.1.0
fpdf2>=2.7.0
